+++ R/btt22-package.R
@@ -1,0 +1,11 @@
+url_ussie <- function() {
+  "https://github.com/rstudio-conf-2022/ussie"
+}
+
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+#' @importFrom rlang %||%
+## usethis namespace: end
+NULL