+++ NEWS.md
@@ -1,4 +1,11 @@
-# btt22 0.0.0.9000
+# btt22 0.0.1
+
+* This package keeps a copy of the {ussie} repo. (#11)
+
+  - Added option `"btt22.source"` to use `"local"` (default) or `"remote"` source for {ussie}. 
+  - For instructors, added unexported function, `btt_update_ussie()`:
+    - puts latest copy of ussie repo, as a `.zip` file, into `/inst/ussie-repo`
+    - bumps the (patch) version of this package
 
 * Imports {devtools}, {diffviewer}, {dplyr}, {engsoccerdata}, {ggplot2}, and {shiny}, so that installing this package will get you all the packages you will need for the course. (#15)